Refactor activities in packages
