+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
 
         <!-- ✅ SplashActivity ist Launcher und braucht exported=true -->
         <activity
-            android:name="at.aau.serg.sdlapp.ui.SplashActivity"
+            android:name="at.aau.serg.sdlapp.ui.activity.SplashActivity"
             android:exported="true"
             android:theme="@style/Theme.SDLApp.Splash"
             android:screenOrientation="landscape">
@@ -32,14 +32,14 @@
 
         <!-- ✅ StartActivity: exported=false (keine externe Aufrufbarkeit nötig) -->
         <activity
-            android:name="at.aau.serg.sdlapp.ui.StartActivity"
+            android:name="at.aau.serg.sdlapp.ui.activity.StartActivity"
             android:exported="false"
             android:theme="@style/Theme.MyApplication"
             android:screenOrientation="landscape" />
 
         <!-- ✅ SettingsActivity -->
         <activity
-            android:name="at.aau.serg.sdlapp.ui.SettingsActivity"
+            android:name="at.aau.serg.sdlapp.ui.activity.SettingsActivity"
             android:exported="false"
             android:label="Einstellungen"
             android:theme="@style/Theme.MyApplication"
@@ -47,14 +47,14 @@
 
         <!-- ✅ BoardActivity -->
         <activity
-            android:name="at.aau.serg.sdlapp.ui.BoardActivity"
+            android:name="at.aau.serg.sdlapp.ui.activity.BoardActivity"
             android:exported="false"
             android:theme="@style/Theme.MyApplication"
             android:screenOrientation="landscape" />
 
         <!-- JobSelectionActivity -->
         <activity
-            android:name="at.aau.serg.sdlapp.ui.JobSelectionActivity"
+            android:name="at.aau.serg.sdlapp.ui.activity.job.JobSelectionActivity"
             android:exported="false"
             android:label="Job Auswahl"
             android:screenOrientation="landscape"
@@ -62,7 +62,7 @@
 
         <!-- PlayerStatsActivity -->
         <activity
-            android:name="at.aau.serg.sdlapp.ui.PlayerStatsActivity"
+            android:name="at.aau.serg.sdlapp.ui.activity.PlayerStatsActivity"
             android:exported="true"
             android:label="Spielerübersicht"
             android:screenOrientation="landscape"
@@ -70,7 +70,7 @@
 
         <!-- HomeScreenActivity -->
         <activity
-            android:name="at.aau.serg.sdlapp.ui.HomeScreenActivity"
+            android:name="at.aau.serg.sdlapp.ui.activity.HomeScreenActivity"
             android:exported="true"
             android:label="Homescreen"
             android:screenOrientation="landscape"
@@ -78,7 +78,7 @@
 
         <!-- LobbyActivity -->
         <activity
-            android:name="at.aau.serg.sdlapp.ui.LobbyActivity"
+            android:name="at.aau.serg.sdlapp.ui.activity.LobbyActivity"
             android:exported="true"
             android:label="Lobby"
             android:screenOrientation="landscape"
@@ -86,7 +86,7 @@
 
         <!-- JobCardActivity for Functionality Test-->
         <activity
-            android:name="at.aau.serg.sdlapp.ui.JobCardActivity"
+            android:name="at.aau.serg.sdlapp.ui.activity.job.JobCardActivity"
             android:exported="false"
             android:label="Jobkarte"
             android:screenOrientation="landscape"
