+++ app/src/main/java/at/aau/serg/sdlapp/ui/activity/WheelActivity.kt
@@ -1,4 +1,4 @@
-package at.aau.serg.sdlapp.ui
+package at.aau.serg.sdlapp.ui.activity
 
 import android.os.Bundle
 import android.util.Log
